Extract module loader
